First running version with app factory
